+++ src/main/java/com/example/humorie/global/config/CorsConfig.java
@@ -1,0 +1,19 @@
+package com.example.humorie.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class CorsConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins("*")
+                .allowedMethods("*")
+                .allowedHeaders("Authorization", "Content-Type", "refresh_token")
+                .maxAge(3600);
+    }
+
+}
