+++ src/main/java/com/example/humorie/global/config/SwaggerConfig.java
@@ -1,0 +1,34 @@
+package com.example.humorie.account.config;
+
+import io.swagger.v3.oas.models.Components;
+import io.swagger.v3.oas.models.security.SecurityRequirement;
+import io.swagger.v3.oas.models.security.SecurityScheme;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+
+    @Bean
+    public OpenAPI openAPI() {
+        Info info = new Info()
+                .title("API 문서")
+                .description("잘못된 부분이나 오류 발생 시 말씀해주세요.");
+
+        SecurityScheme bearerAuth = new SecurityScheme()
+                .type(SecurityScheme.Type.APIKEY)
+                .in(SecurityScheme.In.HEADER)
+                .name("Authorization");
+
+        SecurityRequirement addSecurityItem = new SecurityRequirement();
+        addSecurityItem.addList("Bearer Token");
+
+        return new OpenAPI()
+                .components(new Components().addSecuritySchemes("Bearer Token", bearerAuth))
+                .addSecurityItem(addSecurityItem)
+                .info(info);
+    }
+
+}
